+++ apps/meteor/ee/app/license/server/index.ts
@@ -1,5 +1,3 @@
 import './settings';
 import './methods';
-import './startup';
-
-export { getStatistics } from './getStatistics';
+// import './startup';
