+++ apps/meteor/ee/server/index.ts
@@ -1,4 +1,5 @@
-import '../app/license/server/index';
+import './models/startup';
+import '../app/license/server';
 import '../app/api-enterprise/server/index';
 import '../app/authorization/server/index';
 import '../app/canned-responses/server/index';
@@ -11,3 +12,5 @@ import './requestSeatsRoute';
 import './configuration/index';
 import './local-services/ldap/service';
 import './methods/getReadReceipts';
+
+export { registerEEBroker } from './startup';
