+++ apps/meteor/ee/server/startup/apps/trialExpiration.ts
@@ -1,10 +1,9 @@
 import { License } from '@rocket.chat/license';
 import { Meteor } from 'meteor/meteor';
 
-import { Apps } from '../../apps';
-
 Meteor.startup(() => {
-	License.onInvalidateLicense(() => {
+	License.onInvalidateLicense(async () => {
+		const { Apps } = await import('../../apps');
 		void Apps.disableApps();
 	});
 });
