+++ apps/meteor/ee/server/startup/index.ts
@@ -11,12 +11,14 @@ import { api } from '@rocket.chat/core-services';
 
 import { isRunningMs } from '../../../server/lib/isRunningMs';
 
-// only starts network broker if running in micro services mode
-if (isRunningMs()) {
-	const { broker } = await import('./broker');
+export const registerEEBroker = async (): Promise<void> => {
+	// only starts network broker if running in micro services mode
+	if (isRunningMs()) {
+		const { broker } = await import('./broker');
 
-	api.setBroker(broker);
-	void api.start();
-} else {
-	require('./presence');
-}
+		api.setBroker(broker);
+		void api.start();
+	} else {
+		require('./presence');
+	}
+};
