--- apps/meteor/imports/startup/server/index.ts
@@ -1,1 +1,0 @@
-import '../../personal-access-tokens/server';
