+++ apps/meteor/server/configureLogLevel.ts
@@ -2,7 +2,9 @@ import type { LogLevelSetting } from '@rocket.chat/logger';
 import { logLevel } from '@rocket.chat/logger';
 import { Settings } from '@rocket.chat/models';
 
-const LogLevel = await Settings.getValueById('Log_Level');
-if (LogLevel) {
-	logLevel.emit('changed', LogLevel as LogLevelSetting);
-}
+export const configureLogLevel = async (): Promise<void> => {
+	const LogLevel = await Settings.getValueById('Log_Level');
+	if (LogLevel) {
+		logLevel.emit('changed', LogLevel as LogLevelSetting);
+	}
+};
