+++ apps/meteor/server/methods/index.ts
@@ -1,0 +1,54 @@
+import './OEmbedCacheCleanup';
+import './addAllUserToRoom';
+import './addRoomLeader';
+import './addRoomModerator';
+import './addRoomOwner';
+import './afterVerifyEmail';
+import './browseChannels';
+import './canAccessRoom';
+import './channelsList';
+import './createDirectMessage';
+import './deleteFileMessage';
+import './deleteUser';
+import './eraseRoom';
+import './getAvatarSuggestion';
+import './getPasswordPolicy';
+import './getRoomById';
+import './getRoomIdByNameOrId';
+import './getRoomNameById';
+import './getSetupWizardParameters';
+import './getTotalChannels';
+import './getUsersOfRoom';
+import './hideRoom';
+import './ignoreUser';
+import './loadHistory';
+import './loadLocale';
+import './loadMissedMessages';
+import './loadNextMessages';
+import './loadSurroundingMessages';
+import './logoutCleanUp';
+import './messageSearch';
+import './muteUserInRoom';
+import './openRoom';
+import './readMessages';
+import './readThreads';
+import './registerUser';
+import './removeRoomLeader';
+import './removeRoomModerator';
+import './removeRoomOwner';
+import './removeUserFromRoom';
+import './reportMessage';
+import './requestDataDownload';
+import './resetAvatar';
+import './roomNameExists';
+import './saveUserPreferences';
+import './saveUserProfile';
+import './sendConfirmationEmail';
+import './sendForgotPasswordEmail';
+import './setAvatarFromService';
+import './setUserActiveStatus';
+import './setUserPassword';
+import './toggleFavorite';
+import './unmuteUserInRoom';
+import './userPresence';
+import './userSetUtcOffset';
