+++ apps/meteor/server/publications/index.ts
@@ -1,0 +1,5 @@
+import './messages';
+import './room';
+import './settings';
+import './spotlight';
+import './subscription';
