+++ apps/meteor/server/routes/index.ts
@@ -1,0 +1,5 @@
+import './avatar';
+import './health';
+import './i18n';
+import './timesync';
+import './userDataDownload';
