+++ apps/meteor/server/services/startup.ts
@@ -31,38 +31,38 @@ import { UploadService } from './upload/service';
 import { VideoConfService } from './video-conference/service';
 import { VoipService } from './voip/service';
 
-const { db } = MongoInternals.defaultRemoteCollectionDriver().mongo;
+export const registerServices = async (): Promise<void> => {
+	const { db } = MongoInternals.defaultRemoteCollectionDriver().mongo;
 
-api.registerService(new AppsEngineService());
-api.registerService(new AnalyticsService());
-api.registerService(new AuthorizationLivechat());
-api.registerService(new BannerService());
-api.registerService(new CalendarService());
-api.registerService(new LDAPService());
-api.registerService(new MediaService());
-api.registerService(new MeteorService());
-api.registerService(new NPSService());
-api.registerService(new RoomService());
-api.registerService(new SAUMonitorService());
-api.registerService(new VoipService(db));
-api.registerService(new OmnichannelService());
-api.registerService(new OmnichannelVoipService());
-api.registerService(new TeamService());
-api.registerService(new UiKitCoreAppService());
-api.registerService(new PushService());
-api.registerService(new DeviceManagementService());
-api.registerService(new VideoConfService());
-api.registerService(new UploadService());
-api.registerService(new MessageService());
-api.registerService(new TranslationService());
-api.registerService(new SettingsService());
-api.registerService(new OmnichannelIntegrationService());
-api.registerService(new ImportService());
-api.registerService(new OmnichannelAnalyticsService());
+	api.registerService(new AppsEngineService());
+	api.registerService(new AnalyticsService());
+	api.registerService(new AuthorizationLivechat());
+	api.registerService(new BannerService());
+	api.registerService(new CalendarService());
+	api.registerService(new LDAPService());
+	api.registerService(new MediaService());
+	api.registerService(new MeteorService());
+	api.registerService(new NPSService());
+	api.registerService(new RoomService());
+	api.registerService(new SAUMonitorService());
+	api.registerService(new VoipService(db));
+	api.registerService(new OmnichannelService());
+	api.registerService(new OmnichannelVoipService());
+	api.registerService(new TeamService());
+	api.registerService(new UiKitCoreAppService());
+	api.registerService(new PushService());
+	api.registerService(new DeviceManagementService());
+	api.registerService(new VideoConfService());
+	api.registerService(new UploadService());
+	api.registerService(new MessageService());
+	api.registerService(new TranslationService());
+	api.registerService(new SettingsService());
+	api.registerService(new OmnichannelIntegrationService());
+	api.registerService(new ImportService());
+	api.registerService(new OmnichannelAnalyticsService());
 
-// if the process is running in micro services mode we don't need to register services that will run separately
-if (!isRunningMs()) {
-	void (async (): Promise<void> => {
+	// if the process is running in micro services mode we don't need to register services that will run separately
+	if (!isRunningMs()) {
 		const { Presence } = await import('@rocket.chat/presence');
 
 		const { Authorization } = await import('./authorization/service');
@@ -75,5 +75,5 @@ if (!isRunningMs()) {
 		// Always register the service and manage licensing inside the service (tbd)
 		api.registerService(new QueueWorker(db, Logger));
 		api.registerService(new OmnichannelTranscript(Logger));
-	})();
-}
+	}
+};
