+++ apps/meteor/server/settings/index.ts
@@ -37,47 +37,43 @@ import { createVConfSettings } from './video-conference';
 import { createWebDavSettings } from './webdav';
 import { createWebRTCSettings } from './webrtc';
 
-async function createSettings() {
-	await Promise.all([
-		createAccountSettings(),
-		createAnalyticsSettings(),
-		createAssetsSettings(),
-		createBotsSettings(),
-		createCallCenterSettings(),
-		createCasSettings(),
-		createCrowdSettings(),
-		createEmojiSettings(),
-		createSoundsSettings(),
-		createDiscussionsSettings(),
-		createEmailSettings(),
-		createE2ESettings(),
-		createFederationSettings(),
-		createFileUploadSettings(),
-		createGeneralSettings(),
-		createIRCSettings(),
-		createLdapSettings(),
-		createLogSettings(),
-		createLayoutSettings(),
-		createMessageSettings(),
-		createMetaSettings(),
-		createMiscSettings(),
-		createMobileSettings(),
-		createOauthSettings(),
-		createOmniSettings(),
-		createOTRSettings(),
-		createPushSettings(),
-		createRateLimitSettings(),
-		createRetentionSettings(),
-		createSetupWSettings(),
-		createSlackBridgeSettings(),
-		createSmarshSettings(),
-		createThreadSettings(),
-		createTroubleshootSettings(),
-		createVConfSettings(),
-		createUserDataSettings(),
-		createWebDavSettings(),
-		createWebRTCSettings(),
-	]);
-}
-
-await createSettings();
+await Promise.all([
+	createAccountSettings(),
+	createAnalyticsSettings(),
+	createAssetsSettings(),
+	createBotsSettings(),
+	createCallCenterSettings(),
+	createCasSettings(),
+	createCrowdSettings(),
+	createEmojiSettings(),
+	createSoundsSettings(),
+	createDiscussionsSettings(),
+	createEmailSettings(),
+	createE2ESettings(),
+	createFederationSettings(),
+	createFileUploadSettings(),
+	createGeneralSettings(),
+	createIRCSettings(),
+	createLdapSettings(),
+	createLogSettings(),
+	createLayoutSettings(),
+	createMessageSettings(),
+	createMetaSettings(),
+	createMiscSettings(),
+	createMobileSettings(),
+	createOauthSettings(),
+	createOmniSettings(),
+	createOTRSettings(),
+	createPushSettings(),
+	createRateLimitSettings(),
+	createRetentionSettings(),
+	createSetupWSettings(),
+	createSlackBridgeSettings(),
+	createSmarshSettings(),
+	createThreadSettings(),
+	createTroubleshootSettings(),
+	createVConfSettings(),
+	createUserDataSettings(),
+	createWebDavSettings(),
+	createWebRTCSettings(),
+]);
